+++ src/modules/rlm_logintime/rlm_logintime.c
@@ -198,7 +198,7 @@ static rlm_rcode_t CC_HINT(nonnull) mod_authorize(void *instance, UNUSED void *t
 	MEM(vp = pair_update_reply(attr_session_timeout, TAG_ANY));
 	vp = fr_pair_find_by_da(request->reply->vps, attr_session_timeout, TAG_ANY);
 	if (vp) {	/* just update... */
-		if (vp->vp_uint32 > left) {
+		if (vp->vp_uint32 > (uint32_t)left) {
 			vp->vp_uint32 = (uint32_t)left;
 			RDEBUG("&reply:Session-vp := %pV", &vp->data);
 		}
